// src/app/interfaces/users.ts
export interface Users  {
    address: string;
    bank: string;
    bankAccountNumber: string;
    cellphone: string;
    createdondate: string;
    department: string;
    email: string;
    firstNames: string;
    firstsale: string;
    idclicker: string;
    identification: string;
    isEmployeeGrupoExito: string;
    lastNames: string;
    lastSale: string;
    municipality: string;
    receiveCommunications: string;
    responseaccountbank: string;
    score: string;
    state: string;
    stateId: number;
    totalCommissions: number;
    totalSales: number;
    typeBankAccount: string;
    userId: number;
    verified: string;
    bankcertificate: string;
    id: number,
    identificationcard1: string;
    identificationcard2: string;
    maxdatebankcertificate: string;
    maxdateidentificationcard1: string;
    maxdateidentificationcard2: string;
    maxdaterut: string;
    maxextensiondatebankcertificate: string;
    maxextensiondateidentificationcard1: string;
    maxextensiondateidentificationcard2: string;
    maxextensiondaterut: string;
    mindatebankcertificate: string;
    mindateidentificationcard1: string;
    mindateidentificationcard2: string;
    mindaterut: string;
    rut: string;
  };